requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
arrow==0.13.1             # via jinja2-time
aspy.yaml==1.2.0          # via pre-commit
astor==0.7.1              # via wemake-python-styleguide
atomicwrites==1.3.0       # via pytest
attrs==18.2.0             # via flake8-bugbear, flake8-eradicate, pytest, wemake-python-styleguide
bandit==1.6.0             # via flake8-bandit
binaryornot==0.4.4
certifi==2019.3.9         # via requests
cfgv==1.6.0               # via pre-commit
chardet==3.0.4            # via binaryornot, requests
click==7.0                # via cookiecutter, pip-tools
cookiecutter==1.6.0
decopatch==1.4.5          # via pytest-cases
docutils==0.14            # via restructuredtext-lint
entrypoints==0.3          # via flake8
eradicate==1.0            # via flake8-eradicate
filelock==3.0.10          # via tox
flake8-annotations-complexity==0.0.2  # via wemake-python-styleguide
flake8-bandit==2.1.1      # via wemake-python-styleguide
flake8-broken-line==0.1.0  # via wemake-python-styleguide
flake8-bugbear==18.8.0    # via wemake-python-styleguide
flake8-builtins==1.4.1    # via wemake-python-styleguide
flake8-coding==1.3.1      # via wemake-python-styleguide
flake8-commas==2.0.0      # via wemake-python-styleguide
flake8-comprehensions==2.1.0  # via wemake-python-styleguide
flake8-debugger==3.1.0    # via wemake-python-styleguide
flake8-django==0.0.4
flake8-docstrings==1.3.0  # via wemake-python-styleguide
flake8-eradicate==0.2.0   # via wemake-python-styleguide
flake8-isort==2.7.0       # via wemake-python-styleguide
flake8-logging-format==0.6.0  # via wemake-python-styleguide
flake8-pep3101==1.2.1     # via wemake-python-styleguide
flake8-polyfill==1.0.2    # via flake8-bandit, flake8-docstrings, pep8-naming
flake8-print==3.1.0       # via wemake-python-styleguide
flake8-quotes==1.0.0      # via wemake-python-styleguide
flake8-string-format==0.2.3  # via wemake-python-styleguide
flake8==3.7.6             # via flake8-bandit, flake8-broken-line, flake8-bugbear, flake8-builtins, flake8-coding, flake8-commas, flake8-comprehensions, flake8-debugger, flake8-django, flake8-docstrings, flake8-eradicate, flake8-isort, flake8-pep3101, flake8-polyfill, flake8-print, flake8-quotes, flake8-string-format, wemake-python-styleguide, yesqa
future==0.17.1            # via cookiecutter
gitdb2==2.0.5             # via gitpython
gitpython==2.1.11         # via bandit
identify==1.4.3           # via pre-commit
idna==2.8                 # via requests
importlib-metadata==0.13  # via pre-commit
importlib-resources==1.0.2  # via pre-commit
isort==4.3.20             # via flake8-isort
jinja2-time==0.2.0        # via cookiecutter
jinja2==2.10.1            # via cookiecutter, jinja2-time
makefun==1.6.8            # via decopatch, pytest-cases
markupsafe==1.1.1         # via jinja2
mccabe==0.6.1             # via flake8
more-itertools==7.0.0     # via pytest
nodeenv==1.3.3            # via pre-commit
pathspec==0.5.9           # via yamllint
pbr==5.2.1                # via stevedore
pep8-naming==0.8.2        # via wemake-python-styleguide
pip-tools==3.8.0
pluggy==0.11.0            # via pytest, tox
poyo==0.4.2               # via cookiecutter
pre-commit==1.16.1
py==1.8.0                 # via pytest, tox
pycodestyle==2.5.0        # via flake8, flake8-bandit, flake8-debugger, flake8-print
pydocstyle==3.0.0         # via flake8-docstrings
pyflakes==2.1.1           # via flake8
pygments==2.4.0
pyhamcrest==1.9.0
pytest-cases==1.6.3
pytest-cookies==0.3.0
pytest==4.5.0
python-dateutil==2.8.0    # via arrow
pyyaml==5.1
requests==2.21.0          # via cookiecutter
restructuredtext-lint==1.3.0
sh==1.12.14
six==1.12.0               # via bandit, cfgv, flake8-print, pip-tools, pre-commit, pydocstyle, pyhamcrest, pytest, python-dateutil, stevedore, tox
smmap2==2.0.5             # via gitdb2
snowballstemmer==1.2.1    # via pydocstyle
stevedore==1.30.1         # via bandit
testfixtures==6.8.2       # via flake8-isort
tokenize-rt==2.2.0        # via yesqa
toml==0.10.0              # via pre-commit, tox
tox==3.9.0
typing-extensions==3.7.2  # via wemake-python-styleguide
urllib3==1.24.3           # via requests
virtualenv==16.5.0        # via pre-commit, tox
wcwidth==0.1.7            # via pytest
wemake-python-styleguide==0.8.1
whichcraft==0.5.2         # via cookiecutter
wrapt==1.11.1             # via pytest-cases
yamllint==1.15.0
yesqa==0.0.11
zipp==0.5.1               # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools==41.0.1        # via flake8-annotations-complexity, pyhamcrest, pytest, tox
